rename splat debug-> splat overlay and extract shader code
