+++ src/shaders/splat-shader.ts
